quad_serial_comm: Change count arg for process_serial_data to size_t
The passed data size can be > 255, which would not work with uint8_t.

+++ quad_serial_comm/include/quad_serial_comm/serial_interface.h
@@ -9,6 +9,6 @@ void encode_serial_msg(const quadrotor_msgs::Serial &msg,
                        std::vector<uint8_t> &serial_data);
 
 void process_serial_data(
-    const uint8_t *data, const uint8_t count,
+    const uint8_t *data, const size_t count,
     boost::function<void(quadrotor_msgs::Serial &)> callback);
 #endif
